[FEAT] 주요 배송 로직 추가(허브간 배송 경로 기록) 및 버그 수정

+++ src/main/java/com/tf4/delivery/dto/request/DeliveryRouteLegPatchRequestDto.java
@@ -11,9 +11,9 @@ public class DeliveryRouteLegPatchRequestDto {
     private UUID departureHubId;
     private UUID arrivalHubId;
     private Double estimatedDistanceKm;
-    private Double estimatedTimeMin;
+    private Long estimatedTimeMin;
     private Double actualDistanceKm;
-    private Double actualTimeMin;
+    private Long actualTimeMin;
     private String status;
     private UUID hubDeliveryAgentId;
 }
